+++ docs/fairlearn.exceptions.rst
@@ -1,0 +1,7 @@
+fairlearn\.exceptions module
+----------------------------
+
+.. automodule:: fairlearn.exceptions
+    :members:
+    :undoc-members:
+    :show-inheritance:
