+++ docs/fairlearn.metrics.rst
@@ -1,9 +1,6 @@
 fairlearn\.metrics package
 ==========================
 
-Module contents
----------------
-
 .. automodule:: fairlearn.metrics
     :members:
     :undoc-members:
