+++ docs/fairlearn.postprocessing.rst
@@ -1,9 +1,6 @@
 fairlearn\.postprocessing package
 =================================
 
-Module contents
----------------
-
 .. automodule:: fairlearn.postprocessing
     :members:
     :undoc-members:
