+++ docs/fairlearn.reductions.rst
@@ -1,9 +1,6 @@
 fairlearn\.reductions package
 =============================
 
-Module contents
----------------
-
 .. automodule:: fairlearn.reductions
     :members:
     :undoc-members:
