+++ docs/fairlearn.rst
@@ -1,34 +1,16 @@
 fairlearn package
 =================
 
-The fairlearn package contains several sub-packages.
+The fairlearn package contains several sub-packages and sub-modules.
 
 Subpackages
 -----------
 
 .. toctree::
 
+    fairlearn.exceptions
     fairlearn.metrics
     fairlearn.postprocessing
     fairlearn.reductions
     fairlearn.widget
 
-Submodules
-----------
-
-fairlearn\.exceptions module
-----------------------------
-
-.. automodule:: fairlearn.exceptions
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: fairlearn
-    :members:
-    :undoc-members:
-    :show-inheritance:
