+++ docs/fairlearn.widget.rst
@@ -1,29 +1,6 @@
 fairlearn\.widget package
 =========================
 
-Submodules
-----------
-
-fairlearn\.widget\.fairlearnDashboard module
---------------------------------------------
-
-.. automodule:: fairlearn.widget.fairlearnDashboard
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-fairlearn\.widget\.fairlearnWidget module
------------------------------------------
-
-.. automodule:: fairlearn.widget.fairlearnWidget
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
 .. automodule:: fairlearn.widget
     :members:
     :undoc-members:
